Add some basic tests for this package
One of these right now fails: The browser locale ("navigator.language") is only used when there is no
locale provided to `setLocale` -- but that never happens.

The tests run inside NodeJS, as we only need to do very limited mocking of a browser environment to get
things working, and it is expected that that won't change much in the future due to the limited purpose
of this package.