Keep track of the "locale" in addition to the "language"
When calling 'init' multiple times the translations might change in a way that we get a better
match for the intended locale. Just using the language will not work in this case, and it seems
reasonable to have an explicit set/get pair.

While there: Document the behaviors of the functions.