Only consider the browser language when the user language is not available
accept-language-negotiator's lookup algorithm seems to prefer exact matches even
with lower quality values, so a user that has configured their browser to one of the supported
languages _exactly_ will not be able to select another language through the inexact matching
algorithm (ie. assuming ['en','fr'] as available and the navigator language 'en' the user will not
be able to try to get 'fr-FR' content).

This may or may not match RFC 4647, but it definitely isn't want we want in this library.